15mini-buildd-workarounds: Improve SHM workaround: Also avoid /dev/shm host 'overmount leaking'.
From 15mini-buildd-workarounds:

'/dev/shm' might be a symlink to '/run/shm' in some
chroots. Depending on the system mini-buildd/sbuild runs on
and what chroot is being build for this may or may not lead to
one or both of these problems:

- shm not mounted properly in build chroot (leading to build errors when shm is used).
- shm mount on *the host* gets overloaded (leading to an shm mount "leaking" on the host).

This workaround

- [in '/etc/schroot/mini-buildd/fstab-generic'] Removes shm form the generic sbuuod mounts (avoids the mount leaking).
- [here] Fixes /dev/shm to a directory in case it's not.
- [here] Always mounts /dev/shm itself.

Debian Bug: https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=728096

When this is fixed (and we have added a verisoned depends on
schroot accordingly), all of the above may be reverted again.