gnupg.py: Modify doctest to use pre-arranged keys (make testsuite not block if entropy is sparse).
To still have the "gen_key" test, using '--debug-quick-random' works for
GPG 2.1, and may be the best option. However, we still need to support
1.4, too.

Instead, this now skips the "gen_key" test, and uses pre-arranged (GPG
1.4) sec|pubring, which are copied to a temporary GPG HOME. This
actually allows us to test the (now static) values of key id and
fingerprint, so this also brings some improvement.

Closes: 834683