chroot.py: Update source chroot's apt line on check().
Previously, the sources.list as prepared at debootstrap time would just
stick with the source chroot. This is fine as long as the archive
associated with the base source stays the same. If this changes however,
the chroot checks would

* use an outdated/slower archive, or even
* break, in case the initial archive is no longer working properly.

[This does not influence package building as we always create our own
sources.list on the fly.]

Now, we update the chroot's sources.list with every check() of a
chroot instance (right before the update/upgrade maintenance).