Fix regression introduced in 1.0.22 (1eeece7): Builds using ccache fail on pre-existing chroots due to missing 'libdir'.
We now guarantee the existence of libdir in builder.py, before the
build. They no longer have anything to do with chroot creation or
removal.

This also means they are not deleted any more at all currently -- data
will stay valid even when chroots get recreated (that's good), but this
may lead to some cruft in the lifetime of an instance.