changes.py: buildrequest: Also exclude .changes (fixes: failed buildlog sometimes not archived) and .buildinfo files.
In case of a

* non-source-only ("deb-ful") user upload
* a build failure on the builder for the user-uploaded arch

the copied arch changes would wrongly be used, subsequently leading to a
wrong internal builder error.

At the user's end, this merely means you can't view that archived build log for
such a failed build (instead gets the confusing internal error exception text).

Not adding *.changes in the buildrequest in the 1st place fixes this. For the future,
adding files from the dsc only seems to be more straightforward.