builder.py: sbuild call: set '--no-arch-all' explicitly (Fixes "false" reprepro errors after successful builds w/ sbuild >= 0.77)
sbuild >= 0.77 does not seem to use this as default any more,
even though it's still described that way in the man page.

Thus, arch=all packages are build on any builders, leading to
reprepro hashsum errors when reprepro tries to include
such an arch=all artifact for the second time. To be precise,
this actually only happens when the resp. build is not
reproducible -- which seems to be always the case for
distributions < stretch due to unreproducible build tools,
afaiu.

Anyway, even w/o that new behaviour, it's more clear/better
to set this explicitly.